feat: moved avatar and added more stories
