+++ storybook/stories.ts
@@ -6,6 +6,6 @@ require('../app/components/text-field/text-field.story')
 require('../app/components/wallpaper/wallpaper.story')
 require('../app/components/icon/icon.story')
 require('../app/components/header/header.story')
+require('../app/components/avatar/Avatar.story')
 require('../app/screens/Dashboard/UserListItem.story')
-require('../app/screens/Dashboard/Avatar.story')
 require('../app/screens/Dashboard/CharacterRolls.story')
